+++ web/src/lib.rs
@@ -1,36 +1,35 @@
-use pasfmt::{make_formatter, FormattingSettings};
+use pasfmt::{make_formatter, FormattingConfig};
 use pasfmt_core::prelude::FileOptions;
 use wasm_bindgen::prelude::*;
 
 #[wasm_bindgen]
 pub struct SettingsWrapper {
-    formatting_settings: FormattingSettings,
+    config: FormattingConfig,
 }
 
 #[wasm_bindgen]
 impl SettingsWrapper {
     #[wasm_bindgen(constructor)]
     pub fn new(settings: String) -> Result<Self, String> {
-        let formatting_settings = toml::from_str(&settings).map_err(|e| e.to_string())?;
-        Ok(SettingsWrapper {
-            formatting_settings,
-        })
+        let config = toml::from_str(&settings).map_err(|e| e.to_string())?;
+        Ok(SettingsWrapper { config })
     }
 
     #[wasm_bindgen]
     pub fn max_line_len(&self) -> u32 {
-        self.formatting_settings.max_line_length()
+        self.config.max_line_length()
     }
 }
 
 #[wasm_bindgen]
 pub fn default_settings_toml() -> String {
-    toml::to_string_pretty(&FormattingSettings::default()).unwrap()
+    toml::to_string_pretty(&FormattingConfig::default()).unwrap()
 }
 
 #[wasm_bindgen]
 pub fn fmt(src: &str, settings: &SettingsWrapper) -> Result<String, String> {
-    let formatter = make_formatter(&settings.formatting_settings).map_err(|e| e.to_string())?;
+    let formatter = make_formatter(&settings.config).map_err(|e| e.to_string())?;
 
     Ok(formatter.format(src, FileOptions::new()))
+
 }