+++ CHANGELOG.md
@@ -33,6 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Parameters to the `Formatter::format` function, allowing per-file options to be passed in
   (e.g. cursors to track).
 - `Reconstructor::reconstruct` to have the signiture of the old `Reconstructor::reconstruct_into_buf`.
+- Renamed `FormattingSettings` to `FormattingConfig`.
 
 ### Added